rename files with dashes
