+++ vending-machine.js
